refactor: 수정한 Input 컴포넌트 적용
