+++ pages/profile/[id]/editor.tsx
@@ -8,8 +8,8 @@ import useUser from '@libs/client/useUser';
 import { makeImageURL } from '@libs/client/utils';
 import Button from '@components/common/Button';
 import ErrorMessage from '@components/common/ErrorMessage';
+import Input from '@components/common/Input';
 import HeadMeta from '@components/headMeta';
-import Input from '@components/input';
 import Layout from '@components/layout';
 import CategoryTabMenu from '@components/profile/categoryTabMenu';
 import ProfileWeb from '@components/profile/profileWeb';
