+++ pages/user/create.tsx
@@ -12,8 +12,8 @@ import FacebookBtn from '@components/auth/facebookBtn';
 import GoogleBtn from '@components/auth/googleBtn';
 import Button from '@components/common/Button';
 import ErrorMessage from '@components/common/ErrorMessage';
+import Input from '@components/common/Input';
 import HeadMeta from '@components/headMeta';
-import Input from '@components/input';
 import InputPassword from '@components/inputPassword';
 import Layout from '@components/layout';
 import APPLE_LOGO from '@public/apple.png';
